Fix compilation on Linux

+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: ohakola <ohakola@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/15 14:42:13 by ohakola           #+#    #+#             */
-/*   Updated: 2020/09/11 16:12:05 by ohakola          ###   ########.fr       */
+/*   Updated: 2021/05/05 12:12:10 by ohakola          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -172,8 +172,7 @@ void			ft_sort_int_tab(int *tab, unsigned int size, int dir);
 long double		ft_abs_long_double(long double nb);
 int				ft_match(char *s1, char *s2);
 void			ft_strrev(char *str);
-size_t			get_num_len(long long unsigned int nb,
-				long long unsigned int base);
+size_t			get_num_len(uint64_t nb, uint64_t base);
 void			ft_capitalize(char *str);
 void			ft_uncapitalize(char *str);
 long double		ft_powl(long double nb, int pow);
